Moved error trace handling into a designated Python file
