+++ share/smack/svcomp/utils.py
@@ -175,6 +175,7 @@ def verify_bpl_svcomp(args):
 def write_error_file(args, status, verifier_output):
   from smack.top import VProperty
   from smack.top import VResult
+  from smack.errtrace import smackdOutput
   #return
   if status is VResult.VERIFIED or status is VResult.UNKNOWN:
     return
@@ -184,7 +185,7 @@ def write_error_file(args, status, verifier_output):
   if args.error_file:
     error = None
     if args.language == 'svcomp':
-      error = smackJsonToXmlGraph(smack.top.smackdOutput(verifier_output), args, hasBug, status)
+      error = smackJsonToXmlGraph(smackdOutput(verifier_output), args, hasBug, status)
     elif hasBug:
       error = smack.top.error_trace(verifier_output, args)
     if error is not None:
